Separate bundle type and functions related to bundle loading
Only moving functions in a new file.
